Improvemnts from review; organize file names (vectors.rs -> tree.rs)
