+++ zebra-chain/src/orchard/tests.rs
@@ -1,4 +1,4 @@
 mod preallocate;
 mod prop;
 mod test_vectors;
-mod vectors;
+mod tree;
