+++ zebra-chain/src/sapling/tests.rs
@@ -1,4 +1,4 @@
 mod preallocate;
 mod prop;
 mod test_vectors;
-mod vectors;
+mod tree;
